Add support for whole word matching
Previously the script would by default provide a match if the typo string was found as a part of a word, leading to false positives. This commit retains the previous behavior when the --substrings option is specified and changes the default behavior of the script to match whole words only.

Each usage has a separate database:

- Whole word matching: data/common_misspellings.txt
  - This file is a mirror of https://en.wikipedia.org/wiki/Wikipedia:Lists_of_common_misspellings/For_machines. Any additions/deletions/corrections should be made to the Wikipedia page and then common_misspellings.txt synced with that page.
- Substring matching: data/typos.txt
  - This file will be manually curated in this repository.